requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.30.6
feedparser==6.0.11
trafilatura==1.7.0
httpx==0.27.2
python-dateutil==2.9.0.post0
